Make content page's edit query parameter into a route
