+++ app/router.js
@@ -29,9 +29,13 @@ Router.map(function() {
     this.route('page-not-found', {path: 'preprints/page-not-found'});
     this.route('submit', {path: 'preprints/submit'});
     this.route('discover', {path: 'preprints/discover'});
-    this.route('content', {path: '/:preprint_id' });
+    this.route('content', {path: '/:preprint_id'}, function() {
+        this.route('edit');
+    });
     this.route('provider', {path: 'preprints/:slug'}, function() {
-        this.route('content', {path: '/:preprint_id'});
+        this.route('content', {path: '/:preprint_id'}, function() {
+            this.route('edit');
+        });
         this.route('discover');
         this.route('submit');
         this.route('page-not-found');
