+++ app/routes/provider/content/edit.js
@@ -1,0 +1,3 @@
+import route from '../../content/edit';
+
+export default route;
