+++ app/routes/provider/content/index.js
@@ -1,0 +1,13 @@
+import route from '../../content/index';
+
+route.reopen({
+    controllerName: 'content.index',
+    renderTemplate(controller, model) {
+        this.render('content.index', {
+            controller,
+            model
+        });
+    }
+});
+
+export default route;
