+++ app/templates/submit.hbs
@@ -1,4 +1,4 @@
-{{title 'Submit'}}
+{{title (if editMode 'Edit' 'Submit')}}
 
 <div class="preprint-submit-header">
     <div class="container">
